rename do.ts to durableObjects.ts
